[change!] Modified API endpoints #106
Changed API endpoints from "/api/v1/*" to "/api/v1/ipam/*"

Closes #106
