+++ openwisp_ipam/urls.py
@@ -1,6 +1,7 @@
 from django.urls import include, path
 
-from .api import get_api_urls, views as ipam_api_views
+from .api import views as ipam_api_views
+from .api.urls import get_api_urls
 
 
 def get_urls(api_views):
@@ -10,7 +11,9 @@ def get_urls(api_views):
         api_views: location for getting API views
     """
     return [
-        path('api/v1/', include((get_api_urls(api_views), 'ipam'), namespace='ipam')),
+        path(
+            'api/v1/ipam', include((get_api_urls(api_views), 'ipam'), namespace='ipam')
+        ),
         path('accounts/', include('openwisp_users.accounts.urls')),
     ]
